refactoring - create new folder and new file for refactoring module and insert data
