--- mock-hours.js
@@ -1,46 +1,0 @@
-const weekHours = [
-  {
-    id: 1,
-    day: "Lundi",
-    openAM: "08:45",
-    closeAM: "12:00",
-    openPM: "14:00",
-    closePM: "18:00",
-  },
-  {
-    id: 2,
-    day: "Mardi",
-    openAM: "08:45",
-    closeAM: "12:00",
-    openPM: "14:00",
-    closePM: "18:00",
-  },
-  {
-    id: 3,
-    day: "Mercredi",
-    openAM: "08:45",
-    closeAM: "12:00",
-    openPM: "14:00",
-    closePM: "18:00",
-  },
-  {
-    id: 4,
-    day: "Jeudi",
-    openAM: "08:45",
-    closeAM: "12:00",
-    openPM: "14:00",
-    closePM: "18:00",
-  },
-  {
-    id: 5,
-    day: "Vendredi",
-    openAM: "08:45",
-    closeAM: "12:00",
-    openPM: "14:00",
-    closePM: "18:00",
-  },
-  { id: 6, day: "Samedi", openAM: "08:45", closeAM: "12:00", close: "Fermé" },
-  { id: 7, day: "Dimanche", close: "Fermé" },
-];
-
-module.exports = weekHours;
