+++ src/models/table.js
@@ -205,62 +205,7 @@ const createImageTableIfNotExists = async () => {
   }
 };
 
-/*------------------------------------------------------------------------------ 
------------------------------------TABLE OPENING_HOURS----------------------------------
-------------------------------------------------------------------------------*/
 
-const checkOpeningHoursTableExists = () => {
-  const checkTableQuery = "SHOW TABLES LIKE 'Opening_hours'";
-
-  return new Promise((resolve, reject) => {
-    db.query(checkTableQuery, (err, results) => {
-      if (err) {
-        console.error("Erreur lors de la vérification de la table :", err);
-        reject(err);
-      } else {
-        const tableExists = results.length > 0;
-        resolve(tableExists);
-      }
-    });
-  });
-};
-
-const createOpening_hoursTableIfNotExists = async () => {
-  try {
-    const tableExists = await checkOpeningHoursTableExists();
-
-    if (!tableExists) {
-      const createTableQuery = `
-        CREATE TABLE Opening_Hours(
-          Opening_Hours_id INT(11) NOT NULL AUTO_INCREMENT PRIMARY KEY, 
-          Days VARCHAR(10) NOT NULL,
-          Open_AM VARCHAR(10) NOT NULL,
-          Close_AM VARCHAR(10) NOT NULL,
-          Open_PM VARCHAR(10) NOT NULL,
-          Close_PM VARCHAR(10) NOT NULL,
-          Home_page_id INT(11),          
-          FOREIGN KEY (Home_page_id) REFERENCES Home_page(Home_page_id)
-        ) ENGINE=INNODB
-      `;
-
-      db.query(createTableQuery, (err, result) => {
-        if (err) {
-          console.error("Erreur lors de la création de la table : ", err);
-          throw err;
-        } else {
-          console.log("Table OpeningHours créée avec succès");
-        }
-      });
-    } else {
-      console.log("La table 'OpeningHours' existe déjà.");
-    }
-  } catch (error) {
-    console.error(
-      "Erreur lors de la vérification de la table OpeningHours :",
-      error
-    );
-  }
-};
 
 /*------------------------------------------------------------------------------ 
 -----------------------------------REVIEW----------------------------------
@@ -496,7 +441,6 @@ module.exports = {
   createHomePageTableIfNotExists,
   createSectionTableIfNotExists,
   createImageTableIfNotExists,
-  createOpening_hoursTableIfNotExists,
   createReviewTableIfNotExists,
   createSecondHandCarTableIfNotExists,
   createContactFormTableIfNotExists,
